src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core
azure-functions
azure-storage-blob
azure-identity
functions-framework==3.*
flask-expects-json
ciso8601
# deepgram-sdk
deepgram-sdk==2.*